+++ FacebookAudienceNetwork/com/mopub/nativeads/FacebookAdRenderer.java
@@ -78,6 +78,9 @@ public class FacebookAdRenderer implements MoPubAdRenderer<FacebookNative.Facebo
         NativeRendererHelper.addTextView(facebookNativeViewHolder.getTextView(), nativeAd.getText());
         NativeRendererHelper.addTextView(facebookNativeViewHolder.getCallToActionView(),
                 nativeAd.getCallToAction());
+        NativeRendererHelper.addTextView(facebookNativeViewHolder.getAdvertiserNameView(),
+                nativeAd.getAdvertiserName());
+
         final RelativeLayout adChoicesContainer =
                 facebookNativeViewHolder.getAdChoicesContainer();
         nativeAd.registerChildViewsForInteraction(facebookNativeViewHolder.getMainView(),
@@ -113,6 +116,8 @@ public class FacebookAdRenderer implements MoPubAdRenderer<FacebookNative.Facebo
         private MediaView mediaView;
         @Nullable
         private AdIconView adIconView;
+        @Nullable
+        private TextView advertiserNameView;
 
         // Use fromViewBinder instead of a constructor
         private FacebookNativeViewHolder() {
@@ -134,6 +139,7 @@ public class FacebookAdRenderer implements MoPubAdRenderer<FacebookNative.Facebo
                     view.findViewById(facebookViewBinder.adChoicesRelativeLayoutId);
             viewHolder.mediaView = view.findViewById(facebookViewBinder.mediaViewId);
             viewHolder.adIconView = view.findViewById(facebookViewBinder.adIconViewId);
+            viewHolder.advertiserNameView = view.findViewById(facebookViewBinder.advertiserNameId);
             return viewHolder;
         }
 
@@ -172,6 +178,10 @@ public class FacebookAdRenderer implements MoPubAdRenderer<FacebookNative.Facebo
             return mediaView;
         }
 
+        @Nullable
+        public TextView getAdvertiserNameView() {
+            return advertiserNameView;
+        }
     }
 
     public static class FacebookViewBinder {
@@ -185,6 +195,7 @@ public class FacebookAdRenderer implements MoPubAdRenderer<FacebookNative.Facebo
         final Map<String, Integer> extras;
         final int mediaViewId;
         final int adIconViewId;
+        final int advertiserNameId;
 
         private FacebookViewBinder(@NonNull final Builder builder) {
             this.layoutId = builder.layoutId;
@@ -195,6 +206,7 @@ public class FacebookAdRenderer implements MoPubAdRenderer<FacebookNative.Facebo
             this.extras = builder.extras;
             this.mediaViewId = builder.mediaViewId;
             this.adIconViewId = builder.adIconViewId;
+            this.advertiserNameId = builder.advertiserNameId;
         }
 
         public static class Builder {
@@ -208,6 +220,7 @@ public class FacebookAdRenderer implements MoPubAdRenderer<FacebookNative.Facebo
             private Map<String, Integer> extras = Collections.emptyMap();
             private int mediaViewId;
             private int adIconViewId;
+            private int advertiserNameId;
 
             public Builder(final int layoutId) {
                 this.layoutId = layoutId;
@@ -259,6 +272,12 @@ public class FacebookAdRenderer implements MoPubAdRenderer<FacebookNative.Facebo
             @NonNull
             public Builder adIconViewId(final int adIconViewId) {
                 this.adIconViewId = adIconViewId;
+                return this;
+            }
+
+            @NonNull
+            public Builder advertiserNameId(final int advertiserNameId) {
+                this.advertiserNameId = advertiserNameId;
                 return this;
             }